Improve the makefile setup (#1597)
* Remove containers when stopping them

This ensures that starting the setup again picks up the new version of
the image during the next run.

* Clean docker setups in components

The components should be pulling the images they launch and not a
different tag. And they should not try to start the shared localstack
container when the image they use is not localstack. Instead, the should
check for the shared container created by the root Makefile

* Clean all docker images of the root Makefile

* Change the root Makefile to be consistent with services

Instead of having `make stop-docker` triggering the `clean` command
(which also triggers the cleaning of sub-services), it is now `clean`
which depends on `stop-docker` of the root repo.
This brings consistency with initialization commands. We now have a
symmetry between `initialize`/`clean` and `start-docker`/`stop-docker`.